+++ admin/words.js
@@ -51,8 +51,20 @@ systemDictionary = {
 		"pl": "Częstotliwość odświeżania",
 		"zh-cn": "刷新率"
 	},
+	"refreshhelper": {
+		"en": "In Seconds - Refresh rate for retrieving data from HUUM device in seconds",
+		"de": "In Sekunden - Aktualisierungsrate zum Abrufen von Daten vom HUUM-Gerät in Sekunden",
+		"ru": "Частота обновления для получения данных с устройства HUUM в секундах",
+		"pt": "Taxa de atualização para recuperar dados do dispositivo HUUM em segundos",
+		"nl": "Vernieuwingsfrequentie voor het binnen enkele seconden ophalen van gegevens van het HUUM-apparaat",
+		"fr": "Taux de rafraîchissement pour récupérer les données de l'appareil HUUM en quelques secondes",
+		"it": "Frequenza di aggiornamento per il recupero dei dati dal dispositivo HUUM in pochi secondi",
+		"es": "Frecuencia de actualización para recuperar datos del dispositivo HUUM en segundos",
+		"pl": "Częstotliwość odświeżania do pobierania danych z urządzenia HUUM w kilka sekund",
+		"zh-cn": "以秒为单位从 HUUM 设备检索数据的刷新率"
+	},
 	"sleep": {
-		"en": "Sleep Refresh rate",
+		"en": "Sleep Refresh Rate",
 		"de": "Aktualisierungsrate",
 		"ru": "Частота обновления",
 		"pt": "Taxa de atualização",
@@ -62,6 +74,18 @@ systemDictionary = {
 		"es": "Frecuencia de actualización",
 		"pl": "Częstotliwość odświeżania",
 		"zh-cn": "刷新率"
+	},
+	"sleephelper": {
+		"en": "In Minutes - Sleep Refresh for Sauna Standby Mode to reduce thte traffic, 0 no traffic",
+		"de": "In Minuten - Sleep Refresh für den Sauna-Standby-Modus, um den Traffic zu reduzieren, Wert = 0 kein lesen",
+		"ru": "В минутах - обновление сна для режима ожидания сауны, чтобы уменьшить трафик, 0 нет трафика",
+		"pt": "Em minutos - Atualização do sono para o modo de espera da sauna para reduzir o tráfego, 0 sem tráfego",
+		"nl": "In minuten - Slaapvernieuwing voor sauna-standby-modus om het verkeer te verminderen, 0 geen verkeer",
+		"fr": "En quelques minutes - Actualisation du sommeil pour le mode veille du sauna pour réduire le trafic, 0 pas de trafic",
+		"it": "In pochi minuti - Ripristino del sonno per la modalità standby della sauna per ridurre il traffico, 0 nessun traffico",
+		"es": "En minutos: actualización del sueño para el modo de espera de sauna para reducir el tráfico, 0 sin tráfico",
+		"pl": "W minutach — odświeżanie snu w trybie czuwania w saunie w celu zmniejszenia ruchu, 0 brak ruchu",
+		"zh-cn": "几分钟内 - 桑拿待机模式的睡眠刷新以减少流量，0 无流量"
 	},
 	"lightpath": {
 		"en": "Path to state for light switching",
@@ -76,8 +100,8 @@ systemDictionary = {
 		"zh-cn": "灯切换的状态路径"
 	},
 	"astrolight": {
-		"en": "Switch On light automatically with switch sauna with astro times",
-		"de": "Licht automatisch einschalten mit Schalter Sauna mit Astrozeiten",
+		"en": "Use Astro",
+		"de": "Astro Funktion nutzen",
 		"ru": "Автоматическое включение света с помощью переключателя сауны с астрономическим временем",
 		"pt": "Acender a luz automaticamente com interruptor sauna com tempos astro",
 		"nl": "Schakel het licht automatisch in met schakelaar sauna met astrotijden",
@@ -87,5 +111,17 @@ systemDictionary = {
 		"pl": "Automatyczne włączanie światła za pomocą przełącznika sauny z czasami astro",
 		"zh-cn": "使用带有天文时代的开关桑拿自动打开灯"
 	},
+	"astrohelper": {
+		"en": "Switch On light automatically using  astro times",
+		"de": "Licht automatisch einschalten wenn es dunkel wird",
+		"ru": "Автоматическое включение света с помощью переключателя сауны с астрономическим временем",
+		"pt": "Acender a luz automaticamente com interruptor sauna com tempos astro",
+		"nl": "Schakel het licht automatisch in met schakelaar sauna met astrotijden",
+		"fr": "Allumer la lumière automatiquement avec interrupteur sauna avec horaires astro",
+		"it": "Accensione automatica della luce con interruttore sauna con orari astronomici",
+		"es": "Encienda la luz automáticamente con el interruptor de sauna con tiempos astronómicos",
+		"pl": "Automatyczne włączanie światła za pomocą przełącznika sauny z czasami astro",
+		"zh-cn": "使用带有天文时代的开关桑拿自动打开灯"
+	}
 
 };